+++ lib/simplecov_mcp/error_handler.rb
@@ -9,9 +9,10 @@ module SimpleCovMcp
   class ErrorHandler
     attr_accessor :error_mode, :logger
 
-    VALID_ERROR_MODES = [:off, :on, :trace].freeze
+    VALID_ERROR_MODES = [:off, :log, :debug].freeze
 
-    def initialize(error_mode: :on, logger: nil)
+    def initialize(error_mode: :log, logger: nil)
+      error_mode = normalize_mode(error_mode)
       unless VALID_ERROR_MODES.include?(error_mode)
         raise ArgumentError, "Invalid error_mode: #{error_mode.inspect}. Valid modes: #{VALID_ERROR_MODES.inspect}"
       end
@@ -25,7 +26,7 @@ module SimpleCovMcp
     end
 
     def show_stack_traces?
-      error_mode == :trace
+      error_mode == :debug
     end
 
     # Handle an error with appropriate logging and re-raising behavior
@@ -160,6 +161,17 @@ module SimpleCovMcp
         "missing method '#{method_name}' on #{object_info}"
       else
         message
+      end
+    end
+
+    def normalize_mode(mode)
+      case mode
+      when :debug, 'debug', 'd', :trace, 'trace', 't'
+        :debug
+      when :log, 'log', 'l', :on, 'on'
+        :log
+      else
+        mode
       end
     end
   end